backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
mysql-connector-python==8.3.0
pydantic==2.6.4
pydantic[email]
yfinance==0.2.37